# [SPARK-34593][SQL] Preserve broadcast nested loop join partitioning and ordering