--- a/sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q61.sf100/explain.txt
+++ b/sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q61.sf100/explain.txt
@@ -1,74 +1,73 @@
 == Physical Plan ==
-* Sort (70)
-+- Exchange (69)
-   +- * Project (68)
-      +- BroadcastNestedLoopJoin Inner BuildRight (67)
-         :- * HashAggregate (47)
-         :  +- Exchange (46)
-         :     +- * HashAggregate (45)
-         :        +- * Project (44)
-         :           +- * BroadcastHashJoin Inner BuildRight (43)
-         :              :- * Project (31)
-         :              :  +- * BroadcastHashJoin Inner BuildRight (30)
-         :              :     :- * Project (24)
-         :              :     :  +- * BroadcastHashJoin Inner BuildRight (23)
-         :              :     :     :- * Project (17)
-         :              :     :     :  +- * BroadcastHashJoin Inner BuildRight (16)
-         :              :     :     :     :- * Project (10)
-         :              :     :     :     :  +- * BroadcastHashJoin Inner BuildRight (9)
-         :              :     :     :     :     :- * Filter (3)
-         :              :     :     :     :     :  +- * ColumnarToRow (2)
-         :              :     :     :     :     :     +- Scan parquet default.store_sales (1)
-         :              :     :     :     :     +- BroadcastExchange (8)
-         :              :     :     :     :        +- * Project (7)
-         :              :     :     :     :           +- * Filter (6)
-         :              :     :     :     :              +- * ColumnarToRow (5)
-         :              :     :     :     :                 +- Scan parquet default.date_dim (4)
-         :              :     :     :     +- BroadcastExchange (15)
-         :              :     :     :        +- * Project (14)
-         :              :     :     :           +- * Filter (13)
-         :              :     :     :              +- * ColumnarToRow (12)
-         :              :     :     :                 +- Scan parquet default.item (11)
-         :              :     :     +- BroadcastExchange (22)
-         :              :     :        +- * Project (21)
-         :              :     :           +- * Filter (20)
-         :              :     :              +- * ColumnarToRow (19)
-         :              :     :                 +- Scan parquet default.promotion (18)
-         :              :     +- BroadcastExchange (29)
-         :              :        +- * Project (28)
-         :              :           +- * Filter (27)
-         :              :              +- * ColumnarToRow (26)
-         :              :                 +- Scan parquet default.store (25)
-         :              +- BroadcastExchange (42)
-         :                 +- * Project (41)
-         :                    +- * BroadcastHashJoin Inner BuildRight (40)
-         :                       :- * Filter (34)
-         :                       :  +- * ColumnarToRow (33)
-         :                       :     +- Scan parquet default.customer (32)
-         :                       +- BroadcastExchange (39)
-         :                          +- * Project (38)
-         :                             +- * Filter (37)
-         :                                +- * ColumnarToRow (36)
-         :                                   +- Scan parquet default.customer_address (35)
-         +- BroadcastExchange (66)
-            +- * HashAggregate (65)
-               +- Exchange (64)
-                  +- * HashAggregate (63)
-                     +- * Project (62)
-                        +- * BroadcastHashJoin Inner BuildRight (61)
-                           :- * Project (59)
-                           :  +- * BroadcastHashJoin Inner BuildRight (58)
-                           :     :- * Project (56)
-                           :     :  +- * BroadcastHashJoin Inner BuildRight (55)
-                           :     :     :- * Project (53)
-                           :     :     :  +- * BroadcastHashJoin Inner BuildRight (52)
-                           :     :     :     :- * Filter (50)
-                           :     :     :     :  +- * ColumnarToRow (49)
-                           :     :     :     :     +- Scan parquet default.store_sales (48)
-                           :     :     :     +- ReusedExchange (51)
-                           :     :     +- ReusedExchange (54)
-                           :     +- ReusedExchange (57)
-                           +- ReusedExchange (60)
+* Sort (69)
++- * Project (68)
+   +- BroadcastNestedLoopJoin Inner BuildRight (67)
+      :- * HashAggregate (47)
+      :  +- Exchange (46)
+      :     +- * HashAggregate (45)
+      :        +- * Project (44)
+      :           +- * BroadcastHashJoin Inner BuildRight (43)
+      :              :- * Project (31)
+      :              :  +- * BroadcastHashJoin Inner BuildRight (30)
+      :              :     :- * Project (24)
+      :              :     :  +- * BroadcastHashJoin Inner BuildRight (23)
+      :              :     :     :- * Project (17)
+      :              :     :     :  +- * BroadcastHashJoin Inner BuildRight (16)
+      :              :     :     :     :- * Project (10)
+      :              :     :     :     :  +- * BroadcastHashJoin Inner BuildRight (9)
+      :              :     :     :     :     :- * Filter (3)
+      :              :     :     :     :     :  +- * ColumnarToRow (2)
+      :              :     :     :     :     :     +- Scan parquet default.store_sales (1)
+      :              :     :     :     :     +- BroadcastExchange (8)
+      :              :     :     :     :        +- * Project (7)
+      :              :     :     :     :           +- * Filter (6)
+      :              :     :     :     :              +- * ColumnarToRow (5)
+      :              :     :     :     :                 +- Scan parquet default.date_dim (4)
+      :              :     :     :     +- BroadcastExchange (15)
+      :              :     :     :        +- * Project (14)
+      :              :     :     :           +- * Filter (13)
+      :              :     :     :              +- * ColumnarToRow (12)
+      :              :     :     :                 +- Scan parquet default.item (11)
+      :              :     :     +- BroadcastExchange (22)
+      :              :     :        +- * Project (21)
+      :              :     :           +- * Filter (20)
+      :              :     :              +- * ColumnarToRow (19)
+      :              :     :                 +- Scan parquet default.promotion (18)
+      :              :     +- BroadcastExchange (29)
+      :              :        +- * Project (28)
+      :              :           +- * Filter (27)
+      :              :              +- * ColumnarToRow (26)
+      :              :                 +- Scan parquet default.store (25)
+      :              +- BroadcastExchange (42)
+      :                 +- * Project (41)
+      :                    +- * BroadcastHashJoin Inner BuildRight (40)
+      :                       :- * Filter (34)
+      :                       :  +- * ColumnarToRow (33)
+      :                       :     +- Scan parquet default.customer (32)
+      :                       +- BroadcastExchange (39)
+      :                          +- * Project (38)
+      :                             +- * Filter (37)
+      :                                +- * ColumnarToRow (36)
+      :                                   +- Scan parquet default.customer_address (35)
+      +- BroadcastExchange (66)
+         +- * HashAggregate (65)
+            +- Exchange (64)
+               +- * HashAggregate (63)
+                  +- * Project (62)
+                     +- * BroadcastHashJoin Inner BuildRight (61)
+                        :- * Project (59)
+                        :  +- * BroadcastHashJoin Inner BuildRight (58)
+                        :     :- * Project (56)
+                        :     :  +- * BroadcastHashJoin Inner BuildRight (55)
+                        :     :     :- * Project (53)
+                        :     :     :  +- * BroadcastHashJoin Inner BuildRight (52)
+                        :     :     :     :- * Filter (50)
+                        :     :     :     :  +- * ColumnarToRow (49)
+                        :     :     :     :     +- Scan parquet default.store_sales (48)
+                        :     :     :     +- ReusedExchange (51)
+                        :     :     +- ReusedExchange (54)
+                        :     +- ReusedExchange (57)
+                        +- ReusedExchange (60)
 
 
 (1) Scan parquet default.store_sales
@@ -121,7 +120,7 @@ Input [7]: [ss_item_sk#1, ss_customer_sk#2, ss_store_sk#3, ss_promo_sk#4, ss_ext
 Output [2]: [i_item_sk#12, i_category#13]
 Batched: true
 Location [not included in comparison]/{warehouse_dir}/item]
-PushedFilters: [IsNotNull(i_category), EqualTo(i_category,Jewelry), IsNotNull(i_item_sk)]
+PushedFilters: [IsNotNull(i_category), EqualTo(i_category,Jewelry                                           ), IsNotNull(i_item_sk)]
 ReadSchema: struct<i_item_sk:int,i_category:string>
 
 (12) ColumnarToRow [codegen id : 2]
@@ -129,7 +128,7 @@ Input [2]: [i_item_sk#12, i_category#13]
 
 (13) Filter [codegen id : 2]
 Input [2]: [i_item_sk#12, i_category#13]
-Condition : ((isnotnull(i_category#13) AND (i_category#13 = Jewelry)) AND isnotnull(i_item_sk#12))
+Condition : ((isnotnull(i_category#13) AND (i_category#13 = Jewelry                                           )) AND isnotnull(i_item_sk#12))
 
 (14) Project [codegen id : 2]
 Output [1]: [i_item_sk#12]
@@ -378,21 +377,17 @@ Join condition: None
 Output [3]: [promotions#33, total#38, CheckOverflow((promote_precision(CheckOverflow((promote_precision(cast(promotions#33 as decimal(15,4))) / promote_precision(cast(total#38 as decimal(15,4)))), DecimalType(35,20), true)) * 100.00000000000000000000), DecimalType(38,19), true) AS ((CAST(promotions AS DECIMAL(15,4)) / CAST(total AS DECIMAL(15,4))) * 100)#40]
 Input [2]: [promotions#33, total#38]
 
-(69) Exchange
-Input [3]: [promotions#33, total#38, ((CAST(promotions AS DECIMAL(15,4)) / CAST(total AS DECIMAL(15,4))) * 100)#40]
-Arguments: rangepartitioning(promotions#33 ASC NULLS FIRST, total#38 ASC NULLS FIRST, 5), ENSURE_REQUIREMENTS, [id=#41]
-
-(70) Sort [codegen id : 17]
+(69) Sort [codegen id : 16]
 Input [3]: [promotions#33, total#38, ((CAST(promotions AS DECIMAL(15,4)) / CAST(total AS DECIMAL(15,4))) * 100)#40]
 Arguments: [promotions#33 ASC NULLS FIRST, total#38 ASC NULLS FIRST], true, 0
 
 ===== Subqueries =====
 
 Subquery:1 Hosting operator id = 1 Hosting Expression = ss_sold_date_sk#6 IN dynamicpruning#7
-ReusedExchange (71)
+ReusedExchange (70)
 
 
-(71) ReusedExchange [Reuses operator id: 8]
+(70) ReusedExchange [Reuses operator id: 8]
 Output [1]: [d_date_sk#8]
 
 Subquery:2 Hosting operator id = 48 Hosting Expression = ss_sold_date_sk#6 IN dynamicpruning#7