--- a/sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q90.sf100/simplified.txt
+++ b/sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q90.sf100/simplified.txt
@@ -1,79 +1,76 @@
-WholeStageCodegen (12)
+WholeStageCodegen (11)
   Sort [am_pm_ratio]
-    InputAdapter
-      Exchange [am_pm_ratio] #1
-        WholeStageCodegen (11)
-          Project [amc,pmc]
-            InputAdapter
-              BroadcastNestedLoopJoin
-                WholeStageCodegen (5)
-                  HashAggregate [count] [count(1),amc,count]
-                    InputAdapter
-                      Exchange #2
-                        WholeStageCodegen (4)
-                          HashAggregate [count,count]
-                            Project
-                              BroadcastHashJoin [ws_sold_time_sk,t_time_sk]
-                                Project [ws_sold_time_sk]
-                                  BroadcastHashJoin [ws_ship_hdemo_sk,hd_demo_sk]
-                                    Project [ws_sold_time_sk,ws_ship_hdemo_sk]
-                                      BroadcastHashJoin [ws_web_page_sk,wp_web_page_sk]
-                                        Project [ws_sold_time_sk,ws_ship_hdemo_sk,ws_web_page_sk]
-                                          Filter [ws_ship_hdemo_sk,ws_sold_time_sk,ws_web_page_sk]
-                                            ColumnarToRow
-                                              InputAdapter
-                                                Scan parquet default.web_sales [ws_sold_time_sk,ws_ship_hdemo_sk,ws_web_page_sk,ws_sold_date_sk]
+    Project [amc,pmc]
+      InputAdapter
+        BroadcastNestedLoopJoin
+          WholeStageCodegen (5)
+            HashAggregate [count] [count(1),amc,count]
+              InputAdapter
+                Exchange #1
+                  WholeStageCodegen (4)
+                    HashAggregate [count,count]
+                      Project
+                        BroadcastHashJoin [ws_sold_time_sk,t_time_sk]
+                          Project [ws_sold_time_sk]
+                            BroadcastHashJoin [ws_ship_hdemo_sk,hd_demo_sk]
+                              Project [ws_sold_time_sk,ws_ship_hdemo_sk]
+                                BroadcastHashJoin [ws_web_page_sk,wp_web_page_sk]
+                                  Project [ws_sold_time_sk,ws_ship_hdemo_sk,ws_web_page_sk]
+                                    Filter [ws_ship_hdemo_sk,ws_sold_time_sk,ws_web_page_sk]
+                                      ColumnarToRow
                                         InputAdapter
-                                          BroadcastExchange #3
-                                            WholeStageCodegen (1)
-                                              Project [wp_web_page_sk]
-                                                Filter [wp_char_count,wp_web_page_sk]
-                                                  ColumnarToRow
-                                                    InputAdapter
-                                                      Scan parquet default.web_page [wp_web_page_sk,wp_char_count]
-                                    InputAdapter
-                                      BroadcastExchange #4
-                                        WholeStageCodegen (2)
-                                          Project [hd_demo_sk]
-                                            Filter [hd_dep_count,hd_demo_sk]
-                                              ColumnarToRow
-                                                InputAdapter
-                                                  Scan parquet default.household_demographics [hd_demo_sk,hd_dep_count]
-                                InputAdapter
-                                  BroadcastExchange #5
-                                    WholeStageCodegen (3)
-                                      Project [t_time_sk]
-                                        Filter [t_hour,t_time_sk]
-                                          ColumnarToRow
-                                            InputAdapter
-                                              Scan parquet default.time_dim [t_time_sk,t_hour]
-                BroadcastExchange #6
-                  WholeStageCodegen (10)
-                    HashAggregate [count] [count(1),pmc,count]
-                      InputAdapter
-                        Exchange #7
-                          WholeStageCodegen (9)
-                            HashAggregate [count,count]
-                              Project
-                                BroadcastHashJoin [ws_sold_time_sk,t_time_sk]
-                                  Project [ws_sold_time_sk]
-                                    BroadcastHashJoin [ws_ship_hdemo_sk,hd_demo_sk]
-                                      Project [ws_sold_time_sk,ws_ship_hdemo_sk]
-                                        BroadcastHashJoin [ws_web_page_sk,wp_web_page_sk]
-                                          Project [ws_sold_time_sk,ws_ship_hdemo_sk,ws_web_page_sk]
-                                            Filter [ws_ship_hdemo_sk,ws_sold_time_sk,ws_web_page_sk]
-                                              ColumnarToRow
-                                                InputAdapter
-                                                  Scan parquet default.web_sales [ws_sold_time_sk,ws_ship_hdemo_sk,ws_web_page_sk,ws_sold_date_sk]
-                                          InputAdapter
-                                            ReusedExchange [wp_web_page_sk] #3
-                                      InputAdapter
-                                        ReusedExchange [hd_demo_sk] #4
+                                          Scan parquet default.web_sales [ws_sold_time_sk,ws_ship_hdemo_sk,ws_web_page_sk,ws_sold_date_sk]
                                   InputAdapter
-                                    BroadcastExchange #8
-                                      WholeStageCodegen (8)
-                                        Project [t_time_sk]
-                                          Filter [t_hour,t_time_sk]
+                                    BroadcastExchange #2
+                                      WholeStageCodegen (1)
+                                        Project [wp_web_page_sk]
+                                          Filter [wp_char_count,wp_web_page_sk]
                                             ColumnarToRow
                                               InputAdapter
-                                                Scan parquet default.time_dim [t_time_sk,t_hour]
+                                                Scan parquet default.web_page [wp_web_page_sk,wp_char_count]
+                              InputAdapter
+                                BroadcastExchange #3
+                                  WholeStageCodegen (2)
+                                    Project [hd_demo_sk]
+                                      Filter [hd_dep_count,hd_demo_sk]
+                                        ColumnarToRow
+                                          InputAdapter
+                                            Scan parquet default.household_demographics [hd_demo_sk,hd_dep_count]
+                          InputAdapter
+                            BroadcastExchange #4
+                              WholeStageCodegen (3)
+                                Project [t_time_sk]
+                                  Filter [t_hour,t_time_sk]
+                                    ColumnarToRow
+                                      InputAdapter
+                                        Scan parquet default.time_dim [t_time_sk,t_hour]
+          BroadcastExchange #5
+            WholeStageCodegen (10)
+              HashAggregate [count] [count(1),pmc,count]
+                InputAdapter
+                  Exchange #6
+                    WholeStageCodegen (9)
+                      HashAggregate [count,count]
+                        Project
+                          BroadcastHashJoin [ws_sold_time_sk,t_time_sk]
+                            Project [ws_sold_time_sk]
+                              BroadcastHashJoin [ws_ship_hdemo_sk,hd_demo_sk]
+                                Project [ws_sold_time_sk,ws_ship_hdemo_sk]
+                                  BroadcastHashJoin [ws_web_page_sk,wp_web_page_sk]
+                                    Project [ws_sold_time_sk,ws_ship_hdemo_sk,ws_web_page_sk]
+                                      Filter [ws_ship_hdemo_sk,ws_sold_time_sk,ws_web_page_sk]
+                                        ColumnarToRow
+                                          InputAdapter
+                                            Scan parquet default.web_sales [ws_sold_time_sk,ws_ship_hdemo_sk,ws_web_page_sk,ws_sold_date_sk]
+                                    InputAdapter
+                                      ReusedExchange [wp_web_page_sk] #2
+                                InputAdapter
+                                  ReusedExchange [hd_demo_sk] #3
+                            InputAdapter
+                              BroadcastExchange #7
+                                WholeStageCodegen (8)
+                                  Project [t_time_sk]
+                                    Filter [t_hour,t_time_sk]
+                                      ColumnarToRow
+                                        InputAdapter
+                                          Scan parquet default.time_dim [t_time_sk,t_hour]