… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q61.sf100/simplified.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q61/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q61/simplified.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q90.sf100/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q90/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q90/simplified.txt