@c21 2021-03-02
It seems that this PR triggers plan change for two TPCDS queries. I am checking the root cause now.