@cloud-fan 2021-03-03
@c21 It's great to see this PR improves the TPCDS plans, but it also reveals that the plan change is hard to check sometime.

Can you help to improve the test framework to include the number of shuffles, reused shuffles, subqueries, and reused subqueries in the golden files?