@c21 2021-03-03
@cloud-fan - thanks for review.

> Can you help to improve the test framework to include the number of shuffles, reused shuffles, subqueries, and reused subqueries in the golden files?

Sure. I agree this it, and I also found plan check during test failure is quite hard for complicated query. I am actually thinking to add information for ALL physical operators.

The added content like:

```
Total number of operators: 5
HashAggregate: 3
Exchange: 2
Scan: 1
```

WDYT?