@cloud-fan 2021-03-03
ALL physical operators may be too many and make the result hard to read. We can define some important ones. What in my mind is: (reused) shuffle, (reused) broadcast, (reused) subqueries, sort/hash/object aggregate, scans.