@c21 2021-03-03
I am fine if we want to start with a subset of operators. Though I feel if the output is sorted on alphabetical order of operator name, it should be easy to pin-point the difference. Create SPARK-34602 for followup, thanks.